Reorganized all code so model files, error log, and db are in a model folder.
