+++ CodingTemplates/CSharp/model/ErrorLog.txt
@@ -1,0 +1,3 @@
+[2020-01-26 19:32:05 -05:00] System.Exception: Test2...
+   at Model.DatabaseFunctions.GetAllUsers() in C:\Users\Rob\source\repos\Basics\codingtemplates\CSharp\model\DatabaseFunctions.cs:line 145
+   at CSharp.Program.HelloUsers() in C:\Users\Rob\source\repos\Basics\codingtemplates\CSharp\Program.cs:line 42
