+++ CodingTemplates/Java/model/DatabaseFunctions.java
@@ -1,0 +1,29 @@
+/**
+ * Handles all calls to the User database.
+ *
+ * Java Runtime Environment (JRE) version used: 1.8.0_241
+ * Java Development Kit (JDK) version used: 1.8.0_212
+ * SQLite Java Database Connectivity (JDBC) API version used: 3.30.1
+ * SQLite version used: 3.30.1
+ *
+ * Styling guide: Google Java Style Guide
+ *     (https://google.github.io/styleguide/javaguide.html) and
+ *     Code Conventions for the Java Programming Language (Oracle: Deprecated)
+ *     (https://www.oracle.com/technetwork/java/javase/documentation/codeconvtoc-136057.html)
+ *
+ * @category  Basics
+ * @package   Java
+ * @author    Rob Garcia <rgarcia@rgprogramming.com>
+ * @license   https://opensource.org/licenses/MIT The MIT License
+ * @link      https://github.com/garciart/Basics
+ * @copyright 2019-2020 Rob Garcia
+ */
+
+package model;
+
+public class DatabaseFunctions {
+
+    public DatabaseFunctions() {
+
+    }
+ }
