+++ CodingTemplates/Java/notes.txt
@@ -1,0 +1,5 @@
+To compile the file, make sure you are in the project directory (i.e., Java) and run the following command:
+javac -d . hello/Hello.java
+
+To run from terminal, make sure you are in the project directory (i.e., Java) and run the following command:
+java hello.Hello
