+++ README.md
@@ -7,9 +7,9 @@
 Some programmers can recite a whole API by heart, but I'm not one of those guys. There was a time that I could tell you exactly where to look for a piece of code in Waite and Prata's C Primer Plus or Peter Norton's Assembly Language Book for the IBM PC. However, due to school, work, and life in general, I, paraphrasing Albus Dumbledore, sometimes find that I have too much junk crammed into my brain housing group to remember what I did five minutes ago. So, this is my repository of basic things to remember and implement: my software developer pensieve.
 
 - [Coding Templates](https://github.com/garciart/Basics/tree/master/CodingTemplates)
-- [System Security Plans](https://github.com/garciart/Basics/tree/master/SecurityPlans)
 - [Source Control Plan](https://github.com/garciart/Basics/tree/master/SourceControlPlan)
 - [Styling Guide](https://github.com/garciart/Basics/tree/master/StylingGuide)
+- [System Security Plans](https://github.com/garciart/Basics/tree/master/SecurityPlans)
 
 ## References
